+++ munch/__init__.py
@@ -521,7 +521,14 @@ try:
         """
         return json.dumps(self, **options)
 
+    def fromJSON(cls, stream, *args, **kwargs):
+        """ Deserializes JSON to Munch or any of its subclasses.
+        """
+        factory = lambda d: cls(*(args + (d,)), **kwargs)
+        return munchify(json.loads(stream), factory=factory)
+
     Munch.toJSON = toJSON
+    Munch.fromJSON = classmethod(fromJSON)
 
 except ImportError:
     pass